+++ src/bin/copyFilesRecursivePromise.ts
@@ -5,11 +5,13 @@ import { copyFileRecursivePromise } from "./copyFileRecursivePromise";
  * Copies several files recursively.
  * @param sources List of source paths of files to be copied.
  */
-export const copyFilesRecursivePromise = (sources: readonly string[] = []) =>
+export const copyFilesRecursivePromise = (
+	sources: ReadonlyArray<string> = []
+) =>
 	/**
 	 * @param targets List of target paths of files to be copied.
 	 */
-	(targets: readonly string[] = []) =>
+	(targets: ReadonlyArray<string> = []) =>
 		Promise.all(
 			arrayZip(sources)(targets).map(([source, target]) =>
 				copyFileRecursivePromise(source)(target)