+++ src/bin/promiseAllSequential.ts
@@ -3,9 +3,10 @@
  * @param promises Array of promises.
  */
 export const promiseAllSequential = <Argument>(argument: Argument) => <Output>(
-	promises: readonly ((promiseArgument: Argument) => Promise<Output>)[] = []
+	promises: ReadonlyArray<(promiseArgument: Argument) => Promise<Output>> = []
 ) =>
 	promises.slice(1).reduce(
+		// eslint-disable-next-line max-params
 		(promise, nextPromise) =>
 			promise?.then(output =>
 				nextPromise(argument).then(data => [...output, data])